+++ .gitignore
@@ -1,7 +1,3 @@
 /.scannerwork
 /.vscode
-<<<<<<< HEAD
 /node_modules
-=======
-/node_modules
->>>>>>> jest
