+++ index.html
@@ -102,7 +102,7 @@
       </div>
       <!-- Compiled and minified JavaScript -->
       <script src="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/js/materialize.min.js"></script>
-      <script src="./script/calcularMedidas.js"></script>
+      <script src="./dist/calcularMedidas-dist.js"></script>
     </main>
   </body>
 </html>
