+++ package.json
@@ -1,62 +1,34 @@
 {
   "name": "proteccion-imprimir-imagen",
   "version": "1.0.0",
-<<<<<<< HEAD
   "description": "Proyecto para imprimir imagenes en hojas A4",
   "main": "webpack.config.js",
-  "dependencies": {
-    "babel-jest": "^26.6.3",
-    "jest": "^26.6.3"
-  },
+  "dependencies": {},
   "devDependencies": {
     "@babel/core": "^7.12.7",
+    "@babel/polyfill": "^7.12.1",
     "@babel/preset-env": "^7.12.7",
+    "babel-jest": "^26.6.3",
     "babel-loader": "^8.2.1",
+    "babel-polyfill": "^6.26.0",
     "html-loader": "^1.3.2",
     "html-webpack-plugin": "^4.5.0",
+    "jest": "^26.6.3",
     "webpack": "^5.6.0",
     "webpack-cli": "^4.2.0"
   },
   "scripts": {
     "test": "jest",
     "build": "webpack --mode production"
-=======
-  "description": "Proyecto para ajustar una imagen a una hoja A4",
-  "main": "index.js",
-  "scripts": {
-    "build": "babel ./script/calcularMedidas.js -o ./script/calcularMedidas-compiled.js",
-    "test": "jest"
->>>>>>> jest
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/samumirandam/proteccion-imprimir-imagen.git"
   },
-<<<<<<< HEAD
   "author": "samumirandam <samu.miranda.martinez@gmail.com>",
-=======
-  "keywords": [
-    "html",
-    "css",
-    "materialize",
-    "javascript",
-    "jest"
-  ],
-  "author": "samumirandam <samuelchampi@gamil.com>",
->>>>>>> jest
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/samumirandam/proteccion-imprimir-imagen/issues"
   },
-<<<<<<< HEAD
   "homepage": "https://github.com/samumirandam/proteccion-imprimir-imagen#readme"
-=======
-  "homepage": "https://github.com/samumirandam/proteccion-imprimir-imagen#readme",
-  "devDependencies": {
-    "@babel/core": "^7.12.7",
-    "@babel/preset-env": "^7.12.7",
-    "babel-jest": "^26.6.3",
-    "jest": "^26.6.3"
-  }
->>>>>>> jest
 }
